[FIXED] Lesson04. Task4_35 - time as double. Task4_63 - fixed weekend day calculation. Task6_8 - removed duplicated pow calculation

+++ lesson04/src/Task4_35.java
@@ -3,7 +3,7 @@ import java.util.Scanner;
 public class Task4_35 {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int time = sc.nextInt();
-        System.out.println(time % 5 > 3 ? "RED" : "GREEN");
+        double time = sc.nextDouble();
+        System.out.println(time % 5.0 > 3.0 ? "RED" : "GREEN");
     }
 }
